model/requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Segmentation Models
segmentation-models-pytorch>=0.3.3

# Data Processing
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0
roboflow>=1.1.0

# Model Conversion
onnx>=1.14.0
onnxruntime>=1.15.0
tf2onnx>=1.15.0
onnx-tf>=1.10.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pyyaml>=6.0
